packages/playwright-healing-python/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "playwright-healing"
version = "1.0.0"
description = "Autonomous selector healing engine for Playwright Python tests with AI-powered analysis"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Shifty AI", email = "opensource@shifty.ai"}
]
maintainers = [
    {name = "Shifty AI", email = "opensource@shifty.ai"}
]
keywords = [
    "playwright",
    "testing",
    "healing",
    "selector",
    "automation",
    "ai",
    "self-healing",
    "test-automation",
    "shifty",
    "python",
    "pytest"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Framework :: Pytest",
]
requires-python = ">=3.8"
dependencies = [
    "playwright>=1.40.0",
    "requests>=2.28.0",
    "python-Levenshtein>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/shifty-ai/shifty"
Documentation = "https://github.com/shifty-ai/shifty/tree/main/packages/playwright-healing-python"
Repository = "https://github.com/shifty-ai/shifty"
Issues = "https://github.com/shifty-ai/shifty/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["playwright_healing*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
